More modularization for Renesas
